% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeTable-methods.R
\docType{methods}
\name{as.data.frame-LifeTable}
\alias{as.data.frame,LifeTable-method}
\alias{as.data.frame-LifeTable}
\alias{as.data.frame.LifeTable}
\title{Coerce a life table to a data frame.}
\usage{
\method{as.data.frame}{LifeTable}(x, row.names = NULL, optional = FALSE,
  stringsAsFactors = default.stringsAsFactors(), responseName = "value",
  ...)

\S4method{as.data.frame}{LifeTable}(x, row.names = NULL, optional = FALSE,
  stringsAsFactors = default.stringsAsFactors(), responseName = "value",
  ...)
}
\arguments{
\item{x}{An \code{\linkS4class{LifeTable}} object.}

\item{row.names}{See \code{\link[base]{as.data.frame}}.}

\item{optional}{See \code{\link[base]{as.data.frame}}.}

\item{stringsAsFactors}{See \code{\link[base]{as.data.frame}}.}

\item{responseName}{The name of the final column of the table.  Defaults to
\code{"value"}}

\item{\dots}{Additional arguments to be passed to or from methods.}
}
\value{
A \code{\link[base]{data.frame}}.
}
\description{
Convert a \code{\linkS4class{LifeTable}} object to a "long form" data frame.
The life table functions that are included in the data frame are controlled
by the \code{\link{showFun}} slot of the life table.  If the life table
has iterations, the use of quantiles to summarise
these iterations is controlled by the \code{\link{showQuantiles}} and
\code{\link{prob}} slots.
}
\details{
\code{as.data.frame} is often called as a step towards exporting
a life table as a .csv file.

\code{as.data.frame} prints the dimensions of the life table in reverse
order, with the \code{"age"} dimension last.  This ordering is different
from the one used by base function \code{\link[base]{as.data.frame.array}},
but, with life tables, is usually what is wanted.
}
\examples{
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
lt <- LifeTable(mx)
lt
as.data.frame(lt)

showFun(lt) <- c("ex", "Tx")
as.data.frame(lt)


## life table with iterations
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
mx <- perturb(1000 * mx, n = 10) / 1000
lt <- LifeTable(mx)
lt
as.data.frame(lt)
showQuantiles(lt) <- FALSE
lt
as.data.frame(lt)
showQuantiles(lt) <- TRUE
prob(lt) <- c(0.1, 0.8)
lt
as.data.frame(lt)
}
\seealso{
Life tables are created by function \code{\link{LifeTable}}.
}

