% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeTable-methods.R
\docType{methods}
\name{collapseIntervals-LifeTable}
\alias{collapseIntervals-LifeTable}
\alias{collapseIntervals,LifeTable,ANY,ANY,ANY,ANY,missing-method}
\alias{collapseIntervals,LifeTable,ANY,ANY,ANY,ANY,Counts-method}
\title{Collapse intervals in a life table.}
\usage{
\S4method{collapseIntervals}{LifeTable,ANY,ANY,ANY,ANY,missing}(object,
  dimension, breaks = NULL, width = NULL, old = NULL, weights, ...)

\S4method{collapseIntervals}{LifeTable,ANY,ANY,ANY,ANY,Counts}(object,
  dimension, breaks = NULL, width = NULL, old = NULL, weights, ...)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Name or index of the dimension where the intervals are
found.}

\item{breaks}{Numeric vector giving the breaks between intervals after the
merging has occurred.}

\item{width}{The length of the intervals after the merging has occurred.}

\item{old}{The labels of the intervals to be merged.}

\item{weights}{Object of class \code{\linkS4class{Counts}}. Optional.}

\item{\dots}{Not currently used.}
}
\value{
A \code{\linkS4class{LifeTable}}
}
\description{
Collapse intervals in an object of class \code{\linkS4class{LifeTable}}.
The intervals are typically age intervals, that is, age groups.
However, other intervals, other types of intervals, such as time periods,
can be collapsed instead.  See the documentation for
function \code{\link[dembase]{collapseIntervals}}
in package \code{dembase} for more information on collapsing intervals.
}
\details{
In typical use, no \code{weights} argument is supplied.  Instead,
\code{collapseIntervals} calculates \code{Lx} values for the life table,
values and uses them as weights.
}
\examples{
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
lt <- LifeTable(mx)
lt
collapseIntervals(lt, dimension = "age", width = 5)
collapseIntervals(lt, dimension = "age", breaks = c(1, 10))
collapseIntervals(lt, dimension = "age", old = c("0", "1-4"))

popn <- CountsOne(c(10, 80, 120, 200),
                  labels = c("0", "1-4", "5-9", "10+"),
                  name = "age")
collapseIntervals(lt, dimension = "age", width = 5, weights = popn)
}
\seealso{
Life tables are created by function \code{\link{LifeTable}}.
}
