% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToMx.R
\name{convertToMx}
\alias{convertToMx}
\title{Convert life table function into mortality rates.}
\usage{
convertToMx(object, from = c("qx", "px", "dx", "lx", "Lx", "Tx", "ex"), ax,
  radix = 1e+05)
}
\arguments{
\item{object}{An object of class
\code{\link[dembase:DemographicArray-class]{DemographicArray}}.}

\item{from}{The name of a life table function.}

\item{ax}{An object of class
\code{\link[dembase:DemographicArray-class]{DemographicArray}}
specifying 'separation factors'.}

\item{radix}{A positive number, defaulting to 100000.}
}
\value{
An object of class
\code{\link[dembase:DemographicArray-class]{Values}}.
}
\description{
Given the column of a life table (eg 'qx'), calculate the implied mortality
rates ('mx').
}
\details{
\code{convertToMx} is needed when calculating a life table from values
other than 'mx', since function \code{\link{LifeTable}} constructs the
life table from 'mx' values.

\code{object} must have a dimension with \code{\link[dembase]{dimtype}}
\code{"age"}.  \code{from} must be one of  \code{"qx"}, \code{"px"},
\code{"dx"}, \code{"lx"}, \code{"Lx"}, \code{"Tx"}, or \code{"ex"}.
See \code{\link{LifeTable}} for definitions of these functions.

Demographers typically obtain the 'lx' column of a life table by
multiplying the probability of surviving to age 'x' by a 'radix', usually
100000. The 'dx', 'Lx', and 'Tx' columns are also scaled by the 'radix'.

See the documentation for \code{\link{LifeTable}} for a discussion of
how to specify separation factor \code{ax}.

Life table functions \code{qx}, \code{px}, \code{dx}, \code{lx},
effectively contain no information about mortality conditions in the final
age group if, as is normally the case, the final age group is open (that
is, has no upper limit.) With \code{qx}, the final value is always
1.  With \code{px} the final value is always 0.  With \code{dx} the
final value is always the radix minus deaths in earlier age groups.  With
\code{lx}, the final value describes mortality up to age \code{x}, but
says nothing about mortality beyond age \code{x}. (In contrast to the
other life table functions, the first value of \code{lx} contains
no information, since it always equals \code{radix}.)

Because life table functions \code{qx}, \code{px}, \code{dx}, and \code{lx},
contain no information on the final age group, when \code{from} is one of
these functions, the return value from \code{convertToMx} drops the final
age group.  See below for an example.
}
\examples{
Lx <- Counts(array(c(1800000, 1650000, 1500000, 900000, 100000,
                     1700000, 1500000, 1400000, 800000,  50000),
                   dim = c(5, 2),
                   dimnames = list(age = c("0-19", "20-39", "40-59",
                                           "60-79", "80+"),
                                   sex = c("Female", "Male"))))
ax <- Values(array(c(8.3, 5),
                   dim = c(1, 2),
                   dimnames = list(age = "0-19",
                                   sex = c("Female", "Male"))))
convertToMx(Lx, from = "Lx", ax = ax)


ax <- ValuesOne(3, labels = "0-19", name = "age")

lx <- CountsOne(c(100000, 90000, 70000, 30000, 5000),
                labels = c(0, 20, 40, 60, 80),
                name = "age")
ax <- ValuesOne(5, labels = "0-19", name = "age")
convertToMx(lx, from = "lx", ax = ax)

qx <- ValuesOne(c(0.1, 0.05, 0.15, 0.25, 1),
                labels = c("0-19", "20-39", "40-59", "60-79", "80+"),
                name = "age")
ax <- ValuesOne(5, labels = "0-19", name = "age")
convertToMx(qx, from = "qx", ax = ax)
}
\references{
Preston S, Heuveline P, Guillot M. 2000. \emph{Demography:
Measuring and Modeling Population Processes}. Blackwell.
}
\seealso{
\code{convertToMx} may be needed to prepare the \code{mx}
argument for function \code{\link{LifeTable}}.
}

