% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demlife-package.R
\docType{package}
\name{dembase}
\alias{dembase}
\alias{dembase-package}
\title{Functions and data structures for life tables.}
\description{
Tools for creating life tables and extracting life table 'functions'
(also known as life table 'columns').  Uses S4 classes and methods,
and depends on package \code{dembase}.
}
\details{
The package is still under development.  Future features will include
\itemize{
  \item Default values for 'ax'.
  \item Plotting methods for life tables.
  \item Multi-decrement life tables.
  \item Mult-state life tables (eventually).
}
}

