% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demlife-package.R
\docType{package}
\name{demlife}
\alias{demlife}
\alias{demlife-package}
\title{Functions and data structures for life tables.}
\description{
Tools for creating life tables and extracting life table functions
(also known as life table columns).  Uses S4 classes and methods,
and depends on package \code{dembase}.
}
\details{
The package is still under development.  Future features will include
\itemize{
  \item Plotting methods for life tables.
  \item Multi-decrement life tables.
  \item Mult-state life tables (eventually).
}
}
