% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/LifeTable-methods.R
\docType{class}
\name{LifeTable-class}
\alias{LifeTable-class}
\alias{show,LifeTable-method}
\title{S4 class for describing a life table.}
\usage{
\S4method{show}{LifeTable}(object)
}
\arguments{
\item{object}{An object of class \code{"LifeTable"}.}
}
\description{
An object of class \code{LifeTable} holds values for 'mx' and 'ax',
plus the life table 'radix' (ie l0). These provide sufficient information
to calculate all life table functions.
}
\details{
At present, objects of class \code{LifeTable} can only handle
single-decrement life tables.  In future they will be extended to
handle multiple-decrement and multi-state life tables.

Objects of class \code{LifeTable} also holds names of the life table
functions that are displayed when the object is printed, that are
included when the object is turned into a data.frame.
}
\section{Slots}{

\describe{
\item{\code{mx}}{An object of class
\code{\link[dembase:DemographicArray-class]{Values}}, holding mortality
rates.}

\item{\code{ax}}{An object of class
\code{\link[dembase:DemographicArray-class]{Values}}, with the same
metadata as \code{mx}, holding separation factors.}

\item{\code{radix}}{A positive number.}

\item{\code{showFun}}{A character vector with names of life table
functions.}

\item{\code{showQuantiles}}{\code{TRUE} or \code{FALSE}.}

\item{\code{showTotal}}{\code{TRUE} or \code{FALSE}.}

\item{\code{prob}}{A numeric vector.}
}}

\seealso{
Objects of class \code{LifeTable} are typically created using
function \code{\link{LifeTable}}.
}
