% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeTable-generators.R
\name{LifeTable}
\alias{LifeTable}
\title{Create a LifeTable object.}
\usage{
LifeTable(mx, ax = NULL, showFun = c("mx", "qx", "dx", "lx", "Lx", "ex"),
  radix = 1e+05, showQuantiles = TRUE, prob = c(0.025, 0.5, 0.975))
}
\arguments{
\item{mx}{An object of class
\code{\link[dembase:DemographicArray-class]{Values}} holding estimated
mortality rates.}

\item{ax}{An object of class 
\code{\link[dembase:DemographicArray-class]{Values}} holding estimated
separation factors. Optional.}

\item{showFun}{A character vector with the names of the functions that
are printed or plotted by default}

\item{radix}{A positive number.  Defaults to 100,000.}

\item{showQuantiles}{Logical.  If \code{TRUE}, the default,
quantiles are shown, rather than iterations.}

\item{prob}{Values used to calculate quantiles.  Passed to function
\code{\link[stats]{quantile}}.}
}
\value{
An object of class \code{\linkS4class{LifeTable}}.
}
\description{
Create an object of class \code{\linkS4class{LifeTable}} holding the
information necessary to create the full suite of life table functions.
}
\details{
For definitions of the life table functions, see the documentation for
\code{\link{lifeTableFun}}.

\code{mx} is a \code{\link[dembase:DemographicArray-class]{Values}}
object holding estimates of mortality rates, disaggregated by

\code{ax} specifies "separation factors", that is, the average amount of time
lived during an age interval by people who die during that interval.
\code{ax} can not include dimensions that \code{object} does not.  However,
\code{object} \emph{can} include dimensions that \code{ax} does not.  If it
does, then values for \code{ax} are assumed to be identical across these
dimensions. For instance, if \code{object} has a \code{"region"} dimension,
but \code{ax} does not, then the same values for \code{ax} are used for all
regions. See below for examples.

When \code{mx} and \code{ax} share a dimension, all values for that
dimension that appear in \code{mx} must also appear in \code{ax}--except
for the age dimension.  If \code{mx} includes an age interval that
\code{ax} does not, then \code{ax} for that value is assumed to be
half the width of the age interval.

If \code{ax} is not supplied, then \code{LifeTable} imputes values.
The imputed values equal half the length of the corresponding age group,
except at ages 0 and 1-4.  If \code{mx} includes rates for age 0,
then \code{LifeTables} obtains \code{ax} values for ages 0 and 1-4
by applyin the formulas provided by Preston et al (2001: Table 3.3).

The arguments \code{showFun}, \code{showQuantiles} and \code{prob}
affect printing, plotting, and
\code{\link[=as.data.frame-LifeTable]{as.data.frame}}, but do not affect
the underlying data.  This means, for instance, that \code{lifeTableFun} can
be produce values for life table functions that are not included in
\code{showFun}.

\code{l0}, the first value of \code{lx} is conventionally set to 100,000.
Alternative values can be supplied via the \code{radix} argument.  Changing
the value of \code{radix} also affects the \code{dx}, \code{Lx} and
\code{Tx} columns.
}
\examples{
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
ax <- ValuesOne(c(0.3, 1, 2.5, 2.5),
                labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
LifeTable(mx = mx, ax = ax)

al <- demdata::afghan.life
al <- Values(al)
mx <- subarray(al,
               subarray = (fun == "mx") & (time == "2001-2005"))
ax <- subarray(al,
               subarray = (fun == "ax") & (time == "2001-2005"))
LifeTable(mx = mx, ax = ax)
}
\seealso{
\code{\link{lifeTableFun}} returns values for a life table
function.  \code{\link[=as.data.frame-LifeTable]{as.data.frame}}
converts a \code{LifeTable} object to a (long form) data.frame,
typically as a step towards exporting the life table to a .csv file.
}

