% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{lifeTableFun}
\alias{lifeTableFun}
\alias{lifeTableFun,LifeTable-method}
\title{Extract a life table function from a life table.}
\usage{
lifeTableFun(object, fun = c("mx", "qx", "px", "dx", "lx", "Lx", "Tx", "ex"))

\S4method{lifeTableFun}{LifeTable}(object, fun = c("mx", "qx", "px", "dx",
  "lx", "Lx", "Tx", "ex"))
}
\arguments{
\item{object}{An object of class \code{\linkS4class{LifeTable}}.}

\item{fun}{The name of a life table function.}
}
\value{
If \code{fun} is \code{"dx"}, \code{"lx"}, \code{"Lx"}, or
\code{"Tx"}, an object of class \code{\link[dembase]{Counts}};
otherwise an object of class \code{\link[dembase]{Values}}.
}
\description{
Obtain values for a life table function from an object of class
\code{\linkS4class{LifeTable}}.
}
\details{
The life table functions are as follows:
\tabular{ll}{
  Function \tab Definition \cr
  \code{mx} \tab Mortality rate for age-group \code{x}. \cr
  \code{qx} \tab Given survival to exact age x, the probability dying
before the end of the age interval. \cr
  \code{px} \tab \code{1 - qx} \cr
  \code{lx} \tab Of \code{radix} births, the number of people expected to
survive to exact age x \code{radix} \cr
  \code{dx} \tab The number of deaths in the age interval, for a cohort of
size \code{radix} at birth. \cr
  \code{Lx} \tab The expected number of person-years lived in an age interval
by a cohort of size \code{radix} at birth. \cr
  \code{Tx} \tab Sum of the \code{Lx} from age \code{x} to the highest age. \cr
  \code{ex} \tab Life expectancy at age \code{x}. \cr
  \code{ax} \tab The 'separation factor' - the average number of years lived
during an age interval by someone who does during that interval. \cr
}

\code{fun} does not need to be one of the functions specified in the
\code{showFun} argument to \code{\link{LifeTable}}.
}
\examples{
al <- demdata::afghan.life
al <- Values(al)
mx <- subarray(al,
               subarray = (fun == "mx") & (time == "2001-2005"))
ax <- subarray(al,
               subarray = (fun == "ax") & (time == "2001-2005"))
lt <- LifeTable(mx = mx, ax = ax)
lt
lifeTableFun(lt, fun = "qx")
lifeTableFun(lt, fun = "Lx")
## Because it is not included in \\code{\\link{showFun}}, \\code{"Tx"}
## is not displayed.  But 'lifeTableFun' will work anyway.
lifeTableFun(lt, fun = "Tx")
}
\seealso{
\code{\link{lifeExpectancy}} is a convenience function for
extracting life expectancy at a particular age.
}

