% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{prob}
\alias{prob}
\alias{prob,LifeTable-method}
\alias{prob<-}
\alias{prob<-,LifeTable-method}
\title{Quantiles used by a life table.}
\usage{
prob(object)

prob(object) <- value

\S4method{prob}{LifeTable}(object)

\S4method{prob}{LifeTable}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{LifeTable}}.}

\item{value}{A numeric vector.}
}
\value{
The extraction function returns a numeric vector,
and the replacement function returns a \code{\linkS4class{LifeTable}} object
with a new value for the \code{prob} slot.
}
\description{
Extract or change the \code{prob} slot of an object of class
\code{\linkS4class{LifeTable}}.  The \code{prob} slot specifies the
quantiles used in displays, in a life table that has a dimension with
\code{\link[dembase]{dimtype}} \code{"iterations"}.
}
\details{
\code{prob} does not affect the underlying data, which are always
stored as iterations, rather than quantiles.

If \code{object} does not have a dimension with dimtype \code{"iterations"},
then \code{prob} has no effect.

The quantiles must be between 0 and 1 (inclusive) and must be ordered
from smallest to largest.
}
\examples{
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
mx <- perturb(1000 * mx, n = 20) / 1000
lt <- LifeTable(mx)
lt
prob(lt)
prob(lt) <- c(0.1, 0.9)
lt
}
\seealso{
Life tables are created using function \code{\link{LifeTable}}.
}

