% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{radix}
\alias{radix}
\alias{radix<-}
\alias{radix,LifeTable-method}
\alias{radix<-,LifeTable-method}
\title{Get or set the radix of a life table.}
\usage{
radix(object)

radix(object) <- value

\S4method{radix}{LifeTable}(object)

\S4method{radix}{LifeTable}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{LifeTable}}.}

\item{value}{A positive number.}
}
\value{
The extraction function returns a number and the replacement
function returns a \code{\linkS4class{LifeTable}} object with a
new radix.
}
\description{
Extract or change the 'radix' of an object of class
\code{\linkS4class{LifeTable}}.  The radix of a life table is the value for
\code{l0}, the first entry in the \code{lx} column.  As well as affecting
\code{lx}, changing the radix affects \code{dx}, \code{Lx}, and \code{Tx}.
}
\details{
\code{radix} must be a positive number, and defaults to
100,000.
}
\examples{
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
lt <- LifeTable(mx)
radix(lt)
radix(lt) <- 1000
lt
}
\seealso{
Life tables are created using function \code{\link{LifeTable}}.
}
