% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{showFun}
\alias{showFun}
\alias{showFun,LifeTable-method}
\alias{showFun<-}
\alias{showFun<-,LifeTable-method}
\title{Get or set the life table functions shown by default by a life table.}
\usage{
showFun(object)

showFun(object) <- value

\S4method{showFun}{LifeTable}(object)

\S4method{showFun}{LifeTable}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{LifeTable}}.}

\item{value}{A character vector with names of life table functions.}
}
\value{
The extraction function returns a character vector and the
replacement function returns a \code{\linkS4class{LifeTable}} object
with a new value for the \code{showFun} slot.
}
\description{
Extract or change the \code{showFun} slot of an object of class
\code{\linkS4class{LifeTable}}.  The \code{showFun} slot controls the way
that life table functions are displayed, or output via functions
such as \code{\link[=as.data.frame-LifeTable]{as.data.frame}}, but does
not affect the underlying data contained in the
\code{\linkS4class{LifeTable}} object.
}
\details{
See the documentation for \code{\link{lifeTableFun}} for a list of the
valid life table functions.
}
\seealso{
Life tables are created using function \code{\link{LifeTable}}.

mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
ax <- ValuesOne(c(0.3, 1, 2.5, 2.5),
                labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
lt <- LifeTable(mx = mx, ax = ax)
lt
showFun(lt)
showFun(lt) <- c("ex", "lx")
lt
}

