% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{showQuantiles}
\alias{showQuantiles}
\alias{showQuantiles,LifeTable-method}
\alias{showQuantiles<-}
\alias{showQuantiles<-,LifeTable-method}
\title{Whether a LifeTable objects shows quantiles.}
\usage{
showQuantiles(object)

showQuantiles(object) <- value

\S4method{showQuantiles}{LifeTable}(object)

\S4method{showQuantiles}{LifeTable}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{LifeTable}}.}

\item{value}{Logical.}
}
\value{
The extraction function returns \code{TRUE} or \code{FALSE},
and the replacement function returns a \code{\linkS4class{LifeTable}} object
with a new value for the \code{showQuantiles} slot.
}
\description{
Extract or change the \code{showQuantiles} slot of an object of class
\code{\linkS4class{LifeTable}}.  The \code{showQuantiles} slot controls
whether all iterations are shown, or just quantiles summarising those
iterations, in a life table that has a dimension with
\code{\link[dembase]{dimtype}} \code{"iterations"}.
}
\details{
\code{showQuantiles} does not affect the underlying data, which are always
stored as iterations, rather than quantiles.

If \code{object} does not have a dimension with dimtype \code{"iterations"},
then \code{showQuantiles} has no effect.
}
\seealso{
Life tables are created using function \code{\link{LifeTable}}.

mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
ax <- ValuesOne(c(0.3, 1, 2.5, 2.5),
                labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
mx <- perturb(1000 * mx, n = 20) / 1000
lt <- LifeTable(mx = mx, ax = ax)
lt
showQuantiles(lt)
showQuantiles(lt) <- FALSE
lt
}

