% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{showTotal}
\alias{showTotal}
\alias{showTotal,LifeTable-method}
\alias{showTotal<-}
\alias{showTotal<-,LifeTable-method}
\title{Whether a LifeTable objects shows a "total" category for sex}
\usage{
showTotal(object)

showTotal(object) <- value

\S4method{showTotal}{LifeTable}(object)

\S4method{showTotal}{LifeTable}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{LifeTable}}.}

\item{value}{Logical.}
}
\value{
The extraction function returns \code{TRUE} or \code{FALSE},
and the replacement function returns a \code{\linkS4class{LifeTable}} object
with a new value for the \code{showTotal} slot.
}
\description{
Extract or change the \code{showTotal} slot of an object of class
\code{\linkS4class{LifeTable}}.  The \code{showTotal} slot controls
whether a "total" category is shown, in addition to "female"
and "male" categories.  It only has an effect if the life table
has a dimension with \code{\link[dembase]{dimtype}} \code{"sex"}.
}
\details{
\code{showTotal} does not affect the underlying data.  A
\code{\linkS4class{LifeTable}} only stores "female" and "male"
categories; totals are generated as needed.
}
\seealso{
Life tables are created using function \code{\link{LifeTable}}.

al <- demdata::afghan.life
al <- Values(al)
mx <- subarray(al,
               subarray = (fun == "mx") & (time == "2001-2005"))
ax <- subarray(al,
               subarray = (fun == "ax") & (time == "2001-2005"))
lt <- LifeTable(mx = mx, ax = ax)
lt
showTotal(lt)
showTotal(lt) <- FALSE
lt
}

