% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/LifeTable-methods.R
\docType{methods}
\name{Sx}
\alias{Sx}
\alias{Sx,LifeTable-method}
\title{Calculate Sx, the probability of surviving into the next age group.}
\usage{
Sx(object, useLabelStart = TRUE)

\S4method{Sx}{LifeTable}(object, useLabelStart = TRUE)
}
\arguments{
\item{object}{a \code{\linkS4class{LifeTable}}.}

\item{useLabelStart}{Logical. If \code{TRUE}, the default, the
Sx values are labelled by age at the start of the period;
if \code{FALSE}, Sx values are labelled by age at the
end of the period.}
}
\value{
An object of class \code{\link[dembase]{Values}}.
}
\description{
Given a \code{\linkS4class{LifeTable}}, calculate the 'Sx', the probability
of surviving into the next age group.  'Sx' is similar to the life table
function 'Px', except that 'Sx' describes survival from one age group to
the next, while 'Px' describes survival from one exact age to the next.
}
\details{
'Sx' differs from other life table functions in that
\itemize{
  \item it requires all age groups (except the last) to have the same width
(see \code{\link[dembase]{hasRegularAgeTime}}), and
  \item given the same set of data, 'Sx' contains one fewer age group.
}

'Sx' is mainly used in models of population dynamics, such as population
projections, though it occasionally appears in life tables.

If \eqn{L_x} is the life table population between exact ages \eqn{x} and
\eqn{x+n}, then \eqn{S_x = L_{x+a} / L_x}, except for the second-to-last
age group, where \eqn{S_x = (L_{x+a} + L_x) / L_x}.  (\eqn{S_x} is not
defined for the final age group.)  See below for an example.
}
\examples{
mx <- ValuesOne(c(0.2, 0.05, 0.1, 0.4),
               labels = c("0", "1-4", "5-9", "10+"),
               name = "age")
lt <- LifeTable(mx)
lt

## calling 'Sx' on a life table with irregular age intervals
## raises an error
\dontrun{
Sx(lt)
}

## collapse the intervals, and try again
lt <- collapseIntervals(lt, dimension = "age", width = 5)
lt
Sx(lt)

## use age at the end of the interval, rather than the start
Sx(lt, useLabelStart = FALSE)
}
\seealso{
To make the age groups in a life table 'regular', in the sense
described in \code{\link[dembase]{hasRegularAgeTime}}, use function
\code{\link[=collapseIntervals-LifeTable]{collapseIntervals}}.
}

